LPS-179592 Extract ScreenNavigationEntry methods from CPDefinitionCPTypeVirtualScreenNavigationCategory to CPDefinitionCPTypeVirtualScreenNavigationEntry, extends CPDefinitionCPTypeVirtualScreenNavigationCategory in it and and return getCategoryKey in getEntryKey method
